[ES-180][fix][application][flash_driver] modify offset type from uint32_t to uint64_t [rootcause]:prevent overflow [side effect]:none

+++ application/fsbl/alfsbl_emmc_raw.h
@@ -29,7 +29,7 @@ typedef struct{
 }RawEmmcParam_t;
 
 uint32_t AlFsbl_EmmcRawInit(void);
-uint32_t AlFsbl_EmmcRawCopy(uint32_t SrcAddress, PTRSIZE DestAddress, uint32_t Length, SecureInfo *pSecureInfo);
+uint32_t AlFsbl_EmmcRawCopy(uint64_t SrcAddress, PTRSIZE DestAddress, uint32_t Length, SecureInfo *pSecureInfo);
 uint32_t AlFsbl_EmmcRawRelease(void);
 
 #ifdef __cplusplus
